% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bw}
\alias{bw}
\title{bw create a Black&White theme for ggplot2 plot}
\usage{
bw(base_size = 12, base_family = "", noGrid = FALSE)
}
\arguments{
\item{base_size}{[integer]: Font size.}

\item{base_family}{[character]: Font family.}

\item{noGrid}{[logical]: A grid in the background of the plot should be drawn.}
}
\value{
A ggplot2 theme object.
}
\description{
bw create a Black&White theme for ggplot2 plot
}

