% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{manhattan create a Manhanttan plot}
\usage{
manhattan(data, chr, position, y, title = "Manhattan plot",
  xlab = "Chromosomes", ylab = "P-Value", sep = 0.02)
}
\arguments{
\item{data}{[data.frame]: A 'data.frame' which include some columns: chromosome, position and pvalues.}

\item{chr}{[character]: Name of the 'chr' (chromosome) column in 'data'.}

\item{position}{[character]: Name of the 'position' column in 'data'.}

\item{y}{[character]: Name of the 'p-value' column in 'data' (or anything you want to plot with a manhattan plot).}

\item{title}{[character]: The text for the title.}

\item{xlab}{[character]: The text for the x axis.}

\item{ylab}{[character]: The text for the y axis.}

\item{sep}{[numeric]: Width of the space between bars.}
}
\value{
A Manhattan plot in ggplot2 format.
}
\description{
manhattan create a Manhanttan plot
}

