% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.class.R
\name{s.class}
\alias{s.class}
\title{s.class create a s.class plot for a dudi object (ade4)}
\usage{
s.class(dfxy, fac, xax = 1, yax = 2, lab.names = "",
  lab.extreme = FALSE, thresh.extreme = 2, drawEllipse = TRUE,
  cellipse = 1, drawSegment = TRUE)
}
\arguments{
\item{dfxy}{[data.frame]: A data frame containing the two columns for the axes}

\item{fac}{[vector(character)]: A factor partitioning the rows of the data frame in classes.}

\item{xax}{[numeric]: The column number of x in ‘dfxy’. Component from 'dudi.pca' object for x axis.}

\item{yax}{[numeric]: The column number of y in ‘dfxy’. Component from 'dudi.pca' object for y axis.}

\item{lab.names}{[vector(character)]: Should labels be printed.}

\item{lab.extreme}{[logical]: Should outlier's labels be printed.}

\item{thresh.extreme}{[numeric]: Outlier's scale.}

\item{drawEllipse}{[logical]: Draw an ellipse to represent variance.}

\item{cellipse}{[numeric]: Size factor for the ellipse.}

\item{drawSegment}{[logical]: Draw segment from center of each groups defined by 'fac'.}
}
\value{
A s.class plot in ggplot2 format.
}
\description{
s.class create a s.class plot for a dudi object (ade4)
}

