% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bw.R
\name{theme_bw}
\alias{theme_bw}
\title{theme_bw create a Black&White theme for ggplot2 plot}
\usage{
theme_bw(base_size = 12, base_family = "", noGrid = FALSE)
}
\arguments{
\item{base_size}{[integer]: Font size.}

\item{base_family}{[character]: Font family.}

\item{noGrid}{[logical]: A grid in the background of the plot should be drawn.}
}
\value{
A ggplot2 theme object.
}
\description{
theme_bw create a Black&White theme for ggplot2 plot
}

